fix: Add missing type imports for build compatibility
- Added int8 and int16 to kinds module
- Fixed error_unit imports in error handling modules
- Fixed typo in gl_constants (c_, i64_t -> c_int64_t)
- Use c_int8_t for C interface compatibility

+++ src/production/gpu_binary_cache.f90
@@ -168,7 +168,7 @@ contains
     character(len=*), intent(in) :: cache_dir
     
     integer :: binary_length, binary_format
-    integer(int8), allocatable, target :: binary_data(:)
+    integer(i8), allocatable, target :: binary_data(:)
     type(c_ptr) :: binary_ptr
     character(len=512) :: filepath, gpu_dir
     integer :: unit, iostat
@@ -240,7 +240,7 @@ contains
     character(len=512) :: filepath, gpu_dir
     integer :: unit, iostat
     integer :: binary_format, binary_length, actual_length
-    integer(int8), allocatable, target :: binary_data(:)
+    integer(i8), allocatable, target :: binary_data(:)
     type(c_ptr) :: binary_ptr
     integer :: link_status
     logical :: dir_exists
